medico backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.3.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
django-countries==7.6.1
django-phonenumber-field==8.1.0
phonenumbers==9.0.9
djangorestframework-simplejwt==5.5.0
requests==2.32.4
psycopg2-binary==2.9.10
